Encapsulate CursorProcessor compilation
This change encapsulates defining and loading of CursorProcessors
within CursorProcessorCompiler. As a result, it allows to introduce
helper classes used by a CursorProcessor.

Additionally, this improves consistency with other compilers used in the
ExpressionCompiler.
