+++ core/trino-main/src/main/java/io/trino/sql/gen/ExpressionCompiler.java
@@ -14,98 +14,45 @@
 package io.trino.sql.gen;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.collect.ImmutableList;
-import com.google.common.util.concurrent.UncheckedExecutionException;
 import com.google.inject.Inject;
-import io.airlift.bytecode.ClassDefinition;
-import io.trino.cache.CacheStatsMBean;
-import io.trino.cache.NonEvictableLoadingCache;
-import io.trino.metadata.FunctionManager;
 import io.trino.operator.project.CursorProcessor;
 import io.trino.operator.project.PageFilter;
 import io.trino.operator.project.PageProcessor;
 import io.trino.operator.project.PageProjection;
-import io.trino.spi.TrinoException;
 import io.trino.spi.connector.DynamicFilter;
 import io.trino.sql.gen.columnar.ColumnarFilterCompiler;
 import io.trino.sql.gen.columnar.DynamicPageFilter;
 import io.trino.sql.gen.columnar.FilterEvaluator;
 import io.trino.sql.gen.columnar.PageFilterEvaluator;
 import io.trino.sql.relational.RowExpression;
-import org.weakref.jmx.Managed;
-import org.weakref.jmx.Nested;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.OptionalInt;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
-import static com.google.common.base.MoreObjects.toStringHelper;
-import static com.google.common.base.Throwables.throwIfInstanceOf;
 import static com.google.common.collect.ImmutableList.toImmutableList;
-import static io.airlift.bytecode.Access.FINAL;
-import static io.airlift.bytecode.Access.PUBLIC;
-import static io.airlift.bytecode.Access.a;
-import static io.airlift.bytecode.ParameterizedType.type;
-import static io.trino.cache.SafeCaches.buildNonEvictableCache;
-import static io.trino.spi.type.BooleanType.BOOLEAN;
-import static io.trino.sql.gen.BytecodeUtils.invoke;
 import static io.trino.sql.gen.columnar.FilterEvaluator.createColumnarFilterEvaluator;
-import static io.trino.sql.relational.Expressions.constant;
-import static io.trino.util.CompilerUtils.defineClass;
-import static io.trino.util.CompilerUtils.makeClassName;
 import static java.util.Objects.requireNonNull;
 
 public class ExpressionCompiler
 {
+    private final CursorProcessorCompiler cursorProcessorCompiler;
     private final PageFunctionCompiler pageFunctionCompiler;
     private final ColumnarFilterCompiler columnarFilterCompiler;
-    private final NonEvictableLoadingCache<CacheKey, Class<? extends CursorProcessor>> cursorProcessors;
-    private final CacheStatsMBean cacheStatsMBean;
 
     @Inject
-    public ExpressionCompiler(FunctionManager functionManager, PageFunctionCompiler pageFunctionCompiler, ColumnarFilterCompiler columnarFilterCompiler)
+    public ExpressionCompiler(CursorProcessorCompiler cursorProcessorCompiler, PageFunctionCompiler pageFunctionCompiler, ColumnarFilterCompiler columnarFilterCompiler)
     {
-        requireNonNull(functionManager, "functionManager is null");
+        this.cursorProcessorCompiler = requireNonNull(cursorProcessorCompiler, "cursorProcessorCompiler is null");
         this.pageFunctionCompiler = requireNonNull(pageFunctionCompiler, "pageFunctionCompiler is null");
         this.columnarFilterCompiler = requireNonNull(columnarFilterCompiler, "columnarFilterCompiler is null");
-        this.cursorProcessors = buildNonEvictableCache(CacheBuilder.newBuilder()
-                        .recordStats()
-                        .maximumSize(1000),
-                CacheLoader.from(key -> compile(key.getFilter(), key.getProjections(), new CursorProcessorCompiler(functionManager), CursorProcessor.class)));
-        this.cacheStatsMBean = new CacheStatsMBean(cursorProcessors);
-    }
-
-    @Managed
-    @Nested
-    public CacheStatsMBean getCursorProcessorCache()
-    {
-        return cacheStatsMBean;
     }
 
     public Supplier<CursorProcessor> compileCursorProcessor(Optional<RowExpression> filter, List<? extends RowExpression> projections, Object uniqueKey)
     {
-        Class<? extends CursorProcessor> cursorProcessor;
-        try {
-            cursorProcessor = cursorProcessors.getUnchecked(new CacheKey(filter, projections, uniqueKey));
-        }
-        catch (UncheckedExecutionException e) {
-            throwIfInstanceOf(e.getCause(), TrinoException.class);
-            throw e;
-        }
-
-        return () -> {
-            try {
-                return cursorProcessor.getConstructor().newInstance();
-            }
-            catch (ReflectiveOperationException e) {
-                throw new RuntimeException(e);
-            }
-        };
+        return cursorProcessorCompiler.compileCursorProcessor(filter, projections, uniqueKey);
     }
 
     public Function<DynamicFilter, PageProcessor> compilePageProcessor(
@@ -156,104 +103,5 @@ public class ExpressionCompiler
     {
         return () -> compilePageProcessor(true, filter, Optional.empty(), projections, Optional.empty(), OptionalInt.of(initialBatchSize))
                 .apply(DynamicFilter.EMPTY);
-    }
-
-    private <T> Class<? extends T> compile(Optional<RowExpression> filter, List<RowExpression> projections, BodyCompiler bodyCompiler, Class<? extends T> superType)
-    {
-        // create filter and project page iterator class
-        return compileProcessor(filter.orElse(constant(true, BOOLEAN)), projections, bodyCompiler, superType);
-    }
-
-    private <T> Class<? extends T> compileProcessor(
-            RowExpression filter,
-            List<RowExpression> projections,
-            BodyCompiler bodyCompiler,
-            Class<? extends T> superType)
-    {
-        ClassDefinition classDefinition = new ClassDefinition(
-                a(PUBLIC, FINAL),
-                makeClassName(superType.getSimpleName()),
-                type(Object.class),
-                type(superType));
-
-        CallSiteBinder callSiteBinder = new CallSiteBinder();
-        bodyCompiler.generateMethods(classDefinition, callSiteBinder, filter, projections);
-
-        //
-        // toString method
-        //
-        generateToString(
-                classDefinition,
-                callSiteBinder,
-                toStringHelper(classDefinition.getType().getJavaClassName())
-                        .add("filter", filter)
-                        .add("projections", projections)
-                        .toString());
-
-        return defineClass(classDefinition, superType, callSiteBinder.getBindings(), getClass().getClassLoader());
-    }
-
-    private static void generateToString(ClassDefinition classDefinition, CallSiteBinder callSiteBinder, String string)
-    {
-        // bind constant via invokedynamic to avoid constant pool issues due to large strings
-        classDefinition.declareMethod(a(PUBLIC), "toString", type(String.class))
-                .getBody()
-                .append(invoke(callSiteBinder.bind(string, String.class), "toString"))
-                .retObject();
-    }
-
-    private static final class CacheKey
-    {
-        private final Optional<RowExpression> filter;
-        private final List<RowExpression> projections;
-        private final Object uniqueKey;
-
-        private CacheKey(Optional<RowExpression> filter, List<? extends RowExpression> projections, Object uniqueKey)
-        {
-            this.filter = filter;
-            this.uniqueKey = uniqueKey;
-            this.projections = ImmutableList.copyOf(projections);
-        }
-
-        private Optional<RowExpression> getFilter()
-        {
-            return filter;
-        }
-
-        private List<RowExpression> getProjections()
-        {
-            return projections;
-        }
-
-        @Override
-        public int hashCode()
-        {
-            return Objects.hash(filter, projections, uniqueKey);
-        }
-
-        @Override
-        public boolean equals(Object obj)
-        {
-            if (this == obj) {
-                return true;
-            }
-            if (obj == null || getClass() != obj.getClass()) {
-                return false;
-            }
-            CacheKey other = (CacheKey) obj;
-            return Objects.equals(this.filter, other.filter) &&
-                    Objects.equals(this.projections, other.projections) &&
-                    Objects.equals(this.uniqueKey, other.uniqueKey);
-        }
-
-        @Override
-        public String toString()
-        {
-            return toStringHelper(this)
-                    .add("filter", filter)
-                    .add("projections", projections)
-                    .add("uniqueKey", uniqueKey)
-                    .toString();
-        }
     }
 }
