+++ core/trino-main/src/test/java/io/trino/operator/TestColumnarPageProcessor.java
@@ -18,6 +18,7 @@ import io.trino.metadata.FunctionManager;
 import io.trino.operator.project.PageProcessor;
 import io.trino.spi.Page;
 import io.trino.spi.type.Type;
+import io.trino.sql.gen.CursorProcessorCompiler;
 import io.trino.sql.gen.ExpressionCompiler;
 import io.trino.sql.gen.PageFunctionCompiler;
 import io.trino.sql.gen.columnar.ColumnarFilterCompiler;
@@ -81,7 +82,7 @@ public class TestColumnarPageProcessor
 
     private PageProcessor newPageProcessor()
     {
-        return new ExpressionCompiler(functionManager, new PageFunctionCompiler(functionManager, 0), new ColumnarFilterCompiler(functionManager, 0))
+        return new ExpressionCompiler(new CursorProcessorCompiler(functionManager), new PageFunctionCompiler(functionManager, 0), new ColumnarFilterCompiler(functionManager, 0))
                 .compilePageProcessor(Optional.empty(), ImmutableList.of(field(0, types.get(0)), field(1, types.get(1))), MAX_BATCH_SIZE).get();
     }
 }
