+++ core/trino-main/src/test/java/io/trino/sql/gen/BenchmarkInCodeGenerator.java
@@ -187,7 +187,7 @@ public class BenchmarkInCodeGenerator
 
             FunctionManager functionManager = createTestingFunctionManager();
             processor = new ExpressionCompiler(
-                    functionManager,
+                    new CursorProcessorCompiler(functionManager),
                     new PageFunctionCompiler(functionManager, 0),
                     new ColumnarFilterCompiler(functionManager, 0))
                     .compilePageProcessor(
